Create a sample counter slice to test if redux is working
