+++ frontend/src/app/store.ts
@@ -1,7 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit'
+import counterReducer from "../features/testingCounter/counterSlice"
 
 export const store = configureStore({
-  reducer: {},
+  reducer: {
+    counter: counterReducer,
+  },
 })
