--- frontend/src/features/counterSlice.ts
@@ -1,30 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-import type { PayloadAction } from "@reduxjs/toolkit";
-
-export interface CounterState {
-    value:number
-}
-
-const initialState: CounterState = {
-    value:0,
-}
-
-export const counterSlice = createSlice({
-    name:"counter",
-    initialState,
-    reducers:{
-        increment: (state) =>{
-            state.value +=1
-        },
-        decrement: (state) =>{
-            state.value -=1
-        },
-        incrementBuAmount: (state, action: PayloadAction<number>) =>{
-            state.value += action.payload
-        }
-
-    }
-})
-
-export const { increment, decrement, incrementBuAmount } = counterSlice.actions
-export default counterSlice.reducer
