+++ frontend/src/pages/home/home.tsx
@@ -1,11 +1,15 @@
-import Dashboard from "../Dashboard"
+import Dashboard from "../Dashboard";
+import { Counter } from "../../features/testingCounter/Counter";
 const Home = () => {
+  return (
+    <>
+      <Dashboard></Dashboard>
+      <div className="bg-teal-800" h-500>
+        <h1>El contador para ver si redux ta funcionando</h1>
+        <Counter />
+      </div>
+    </>
+  );
+};
 
-    return (
-        <>
-        <Dashboard></Dashboard>
-        </>
-    )
-}
-
-export default Home
+export default Home;
